finalproject/requirements.txt
Django==5.2.5
djangorestframework==3.15.2
django-allauth==65.0.2
django-cors-headers==4.4.0
qrcode==8.0
Pillow==10.4.0
requests==2.32.3